requirements.txt
flask==3.0.3
dash==2.17.1
plotly==5.22.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0
rank-bm25==0.2.2
transformers==4.42.4
arxiv==2.1.3
numpy==1.26.4
gunicorn==22.0.0
